export interfaces and meeting namespace
